updated driver library to be standard format
Update the Arduino driver  to be a standard library format so that that it can
be published by the Arduino Library Manager.

Also address some compilation warnings

+++ drivers/arduino/RGB_LED_Matrix_Lib/src/LEDMatrixBits.h
@@ -49,8 +49,8 @@ public:
 
 	void reset(void);
 	
-	int rows(void) const		{ return _rows; }
-	int columns(void) const		{ return _columns; }
+	size_t rows(void) const			{ return _rows; }
+	size_t columns(void) const		{ return _columns; }
 	
 	void setAllOff(void);
 	void setColumnControlBit( size_t row, size_t column, bool isOn );
